Move instance tests to package
